Move git util to services
