+++ services/git/util.go
@@ -1,7 +1,7 @@
 // Copyright 2021 The Gitea Authors. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-package utils
+package git
 
 import (
 	gocontext "context"
